Keypoint detector finished, need to program the feature matching alogorithm and Ransac from scratch

+++ src/filters.h
@@ -4,7 +4,7 @@
 #include <stdlib.h>
 #include <iostream>
 
-cv::Mat create1DGaussKernel(float std_dev);
+cv::Mat create1DGaussKernel(double std_dev);
 
-cv::Mat create1DDerivateGaussKernel(float std_dev);
+cv::Mat create1DDerivateGaussKernel(double std_dev);
